requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
google-generativeai==0.3.2
langchain-google-genai==0.0.11
langchain==0.1.11
pypdf==3.17.1
python-multipart==0.0.6